+++ src/application/services/CourseService.ts
@@ -5,6 +5,7 @@ import { ICourseRepository } from "../interfaces/IRepositories/ICourseRepository
 import { ICourseService } from "../interfaces/IServices/ICourseService";
 import { ICategoryService } from "../interfaces/IServices/ICategoryService";
 import { CreateCourse, UpdateCourse } from "../inputs/courseInput";
+import { ExtendedUser } from "../types/ExtendedUser";
 import { TransactionType } from "../types/TransactionType";
 import APIError from "../../presentation/errorHandlers/APIError";
 import HttpStatusCode from "../../presentation/enums/HTTPStatusCode";
@@ -13,7 +14,7 @@ import HttpStatusCode from "../../presentation/enums/HTTPStatusCode";
 export class CourseService implements ICourseService {
 	constructor(@inject('ICourseRepository') private courseRepository: ICourseRepository, @inject('ICategoryService') private categoryService: ICategoryService) {}
 
-	async isTrueTopic(id: number): Promise<boolean> {
+	private async isTrueTopic(id: number): Promise<boolean> {
 		const topic = await this.categoryService.findUnique({
 			where: {
 				id
@@ -25,9 +26,27 @@ export class CourseService implements ICourseService {
 		return (topic && topic.type === 'TOPIC') ? true : false;
 	};
 
+	async isResourceBelongsToCurrentUser(courseId: number, user: ExtendedUser): Promise<boolean> {
+		if(!user.roles.includes('Instructor')) {
+			return true;
+		}
+		const course = await this.courseRepository.findFirst({
+			where: {
+				id: courseId,
+				instructor: {
+					userId: user.id
+				}
+			},
+			select: {
+				id: true
+			}
+		});
+		return course ? true : false;
+	}
+
   aggregate(args: Prisma.CourseAggregateArgs): Promise<Prisma.GetCourseAggregateType<Prisma.CourseAggregateArgs>> {
     return this.courseRepository.aggregate(args);
-  }
+  };
 
 	count(args: Prisma.CourseCountArgs): Promise<number> {
 		return this.courseRepository.count(args);
@@ -81,10 +100,13 @@ export class CourseService implements ICourseService {
 	};
 
 	async update(args: {data: UpdateCourse, select?: Prisma.CourseSelect, include?: Prisma.CourseInclude}, transaction?: TransactionType): Promise<Course> {
-    const {id, title, shortDescription, description, language, level, imageCover, requirements, courseTeachings, price, discountPercentage, hours, lectures, articles, quizzes, isApproved, isDraft, sections: sections, topicId} = args.data;
+    const {id, title, shortDescription, description, language, level, imageCover, requirements, courseTeachings, price, discountPercentage, hours, lectures, articles, quizzes, isApproved, isDraft, sections: sections, topicId, user} = args.data;
 		const slug = title ? slugify(title, {lower: true, trim: true}) : undefined;
 		if(topicId && !await this.isTrueTopic(topicId)) {
 			throw new APIError("This topic may be not exist or may be exist but not a topic", HttpStatusCode.BadRequest);
+		}
+		if(user && !await this.isResourceBelongsToCurrentUser(id, user as any)) {
+			throw new APIError('This course is not yours', HttpStatusCode.Forbidden);
 		}
 		return this.courseRepository.update({
 			where: {
@@ -132,7 +154,11 @@ export class CourseService implements ICourseService {
 		}, transaction);
 	};
 
-	delete(id: number, transaction?: TransactionType): Promise<Course> {
+	async delete(args: {id: number, user: ExtendedUser}, transaction?: TransactionType): Promise<Course> {
+		const {id, user} = args;
+		if(!await this.isResourceBelongsToCurrentUser(id, user)) {
+			throw new APIError('This course is not yours', HttpStatusCode.Forbidden);
+		}
 		return this.courseRepository.delete(id, transaction);
 	};
 }
